mcp_servers/hubspot/pyproject.toml
[project]
name = "hubspot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "hubspot-api-client>=11.1.0",
    "mcp[cli]>=1.6.0",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.1",
]
